It turns out that YCPValue, the reference to the real data YCPValueRep, contains TWO pointers: one to the data and other to the virtual method table which is *unused*. It has just a virtual destructor for YCPElement and virtual mem_size.
I made the destructor non-virtual and made mem_size depend on
#ifdef D_MEMUSAGE. That makes YCPValue only ONE pointer. I hope that
besides the memory savings, gcc could optimize it speed-wise too.

But there's a catch. It is binary incompatible in a way that
everything linking with libycp notices -> libycp.so.3

Buy my patch and lose 12% of heap in 3 minutes!
Resident Heap of y2base lan ncurses down to 7584K from 8628K

svn path=/trunk/core/; revision=32989