Fixed Y2DISABLELANGUAGEPLUGINS usage; tests would always succeed (#24).
Use the flag in all tests. Not DRY, but that will be fixed in a planned
refactoring of the *.exp harnesses.

The form that was used before (RUNTEST = Y2DISABLELANGUAGEPLUGINS=1 runtest)
worked only since automake 1.13; but openSUSE 12.3 has 1.12.x and there
the result was that the test failure was ignored and
  WARNING: could not find 'runtest'
was printed after test summary.