Fix YCP symbol comparison with GCC 5 (boo#914255).
This is because of a change in libstdc++: "A new implementation
of std::string is enabled by default, using the small string
optimization instead of copy-on-write reference counting."
https://gcc.gnu.org/gcc-5/changes.html

The priority of the fix is to be ABI compatible with SLE 12.
Therefore we only change the implementation of operator ==;
which makes it slower than an alternative fix involving
pointers, but still acceptable for our use case.