IniAgent: Fix reading files by glob in changed root
It is visible e.g. in .network agent which use multifile functionality.
Buggy behavior was to do first glob and then on result of glob using
changed root behavior. It is fixed by changing to target root whole glob
expression. Test for fix will live in yast2-network to not introduce
rspec dependency in yast core.